% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{latex2exp_examples}
\alias{latex2exp_examples}
\title{latex2exp Examples}
\usage{
latex2exp_examples(cex = 1)
}
\arguments{
\item{cex}{Multiplier for font size}
}
\description{
Plots a number of example LaTeX string, as parsed 
by \code{\link{TeX}}.
}
