% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{latex2exp_supported}
\alias{latex2exp_supported}
\title{Returns the list of supported LaTeX commands.}
\usage{
latex2exp_supported(show = FALSE, ...)
}
\arguments{
\item{show}{Show a searchable table of symbols}

\item{...}{Other parameters (not used)}
}
\value{
A data frame containing a table of supported LaTeX commands.
}
\description{
If \code{show} is TRUE, also show a searchable table of symbols.
}
