% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.latexexpression}
\alias{print.latexexpression}
\title{Print an expression returned by TeX()}
\usage{
\method{print}{latexexpression}(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Ignored}
}
\description{
Prints out the plotmath expression generated by \code{\link{TeX}}
and the original TeX string.
}
