% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex2exp.R
\name{print.latextoken}
\alias{print.latextoken}
\title{Prints out a parsed LaTeX object, as returned by latex2exp(..., output='ast').}
\usage{
\method{print}{latextoken}(x, ...)
}
\arguments{
\item{x}{The latex2exp object.}

\item{...}{(ignored)}
}
\description{
Prints out a parsed LaTeX object, as returned by latex2exp(..., output='ast').
}

