% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex2exp.R
\name{toString.latextoken}
\alias{toString.latextoken}
\title{Converts a token created by latex2exp to a string, later to be parsed into an expression (for internal use).}
\usage{
\method{toString}{latextoken}(x, ...)
}
\arguments{
\item{x}{The latex2exp token}

\item{...}{Additional arguments (ignored)}
}
\value{
A string
}
\description{
Converts a token created by latex2exp to a string, later to be parsed into an expression (for internal use).
}

