% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{validate_input}
\alias{validate_input}
\title{Validates the input LaTeX string}
\usage{
validate_input(latex_string)
}
\arguments{
\item{latex_string}{Input LaTeX string}
}
\description{
Checks for common issues in the LaTeX string, like
unmatched braces.
}
\details{
Also, warns if any of the less common special characters
are present, indicating that perhaps the user accidentally forgot
to escape backslashes.
}
