% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex2exp.R
\name{latex2exp_supported}
\alias{latex2exp_supported}
\title{Returns a list of all supported LaTeX symbols and expressions that can be converted with \code{\link{latex2exp}}.}
\usage{
latex2exp_supported(plot = FALSE)
}
\arguments{
\item{plot}{whether to plot the table (FALSE by default)}
}
\value{
a character vector of supported LaTeX expressions
}
\description{
Returns a list of all supported LaTeX symbols and expressions that can be converted with \code{\link{latex2exp}}.
}

