% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latex2exp.R
\name{plot.expression}
\alias{plot.expression}
\title{Plots an expression on the current graphical device.}
\usage{
\method{plot}{expression}(e, ...)
}
\arguments{
\item{e}{A \code{\link{plotmath}} expression.}

\item{...}{Parameters to be passed to the \code{\link{text}} function.}
}
\description{
Plots an expression on the current graphical device.
}

