% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.latextoken2}
\alias{print.latextoken2}
\title{Prints out a parsed LaTeX object, as returned by TeX(..., output='ast').
This is primarily used for debugging.}
\usage{
\method{print}{latextoken2}(token, depth = 0)
}
\arguments{
\item{x}{The object}

\item{...}{(ignored)}
}
\description{
Prints out a parsed LaTeX object, as returned by TeX(..., output='ast').
This is primarily used for debugging.
}
